move def to black/nodes
