+++ src/blib2to3/pytree.py
@@ -28,12 +28,7 @@ from typing import (
     Iterable,
 )
 
-if sys.version_info >= (3, 10):
-    from typing import TypeGuard
-else:
-    from typing_extensions import TypeGuard
 from blib2to3.pgen2.grammar import Grammar
-from blib2to3.pgen2 import token
 
 __author__ = "Guido van Rossum <guido@python.org>"
 
@@ -986,19 +981,3 @@ def generate_matches(
                     r.update(r0)
                     r.update(r1)
                     yield c0 + c1, r
-
-
-def is_name_token(nl: NL) -> TypeGuard[Leaf]:
-    return nl.type == token.NAME
-
-
-def is_lpar_token(nl: NL) -> TypeGuard[Leaf]:
-    return nl.type == token.LPAR
-
-
-def is_rpar_token(nl: NL) -> TypeGuard[Leaf]:
-    return nl.type == token.RPAR
-
-
-def is_string_token(nl: NL) -> TypeGuard[Leaf]:
-    return nl.type == token.STRING
